test: vitests for facility form
